+++ R/track_methods.R
@@ -89,56 +89,6 @@ centroid.track_xy <- function(x, spatial = FALSE, ...) {
 
 
 
-#' Get bounding box of a track.
-#' @template track_xy_star
-#' @param spatial `[logical(1)=FALSE]` \cr Whether or not to return a `SpatialPolygons`-object or not.
-#' @param buffer `[numeric(0)=NULL]{NULL, >0}` \cr An optional buffer of the bounding box.
-#' @template dots_none
-#' @name bbox
-#' @export
-#' @examples
-#' data(deer)
-#' bbox(deer)
-#' bbox(deer, buffer = 100)
-
-bbox <- function(x, ...) {
-  UseMethod("bbox", x)
-}
-
-#' @export
-#' @rdname bbox
-bbox.track_xy <- function(x, spatial = TRUE, buffer = NULL, sf = FALSE, ...) {
-
-  if (sf && !spatial) {
-    warning("spatial has precedence over sf.")
-  }
-
-  # bounds
-  bbx <- c(min(x$x_), max(x$x_), min(x$y_), max(x$y_))
-
-  if (!is.null(buffer)) {
-    bbx <- bbx + buffer * c(-1, 1, -1, 1)
-  }
-
-  coords <- cbind(
-    x = c(bbx[1], bbx[1], bbx[2], bbx[2], bbx[1]),
-    y = c(bbx[3], bbx[4], bbx[4], bbx[3], bbx[3])
-  )
-
-  p <- sf::st_polygon(list(coords))
-
-  if (spatial) {
-    if (sf) {
-      p
-    } else {
-      sf::as_Spatial(sf::st_geometry(p))
-    }
- } else {
-    sf::st_bbox(p)
-  }
-}
-
-
 #' @export
 points.track_xy <- function(x, ...) {
   graphics::points(x[, c("x_", "y_")], ...)